# backend/api/routers/interpretability.py
"""API router for interpretability features (similarity maps)."""

import asyncio
import logging
from typing import List, Optional

import requests
from api.dependencies import get_colpali_client
from api.models import SimilarityMapRequest, SimilarityMapResponse, TokenInfo, SimilarityMapResult
from domain.errors import ServiceUnavailableError
from fastapi import APIRouter, HTTPException

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/interpretability", tags=["interpretability"])


@router.post("/similarity-map", response_model=SimilarityMapResponse)
async def generate_similarity_map(request: SimilarityMapRequest):
    """Generate interpretability similarity maps for query tokens on an image.

    This endpoint creates visual overlays showing which parts of the image
    are most similar to each query token, enabling interpretability analysis.

    The image is fetched from the provided URL (typically from MinIO storage),
    and similarity maps are generated by the ColPali service.

    Args:
        request: SimilarityMapRequest containing image_url, query, and optional parameters

    Returns:
        SimilarityMapResponse with token info and base64-encoded similarity map images
    """
    try:
        colpali_client = get_colpali_client()
    except Exception as e:
        logger.error("ColPali client not available", extra={"error": str(e)})
        raise HTTPException(
            status_code=503,
            detail="ColPali service unavailable",
        )

    # Fetch the image from the URL
    try:
        logger.debug(
            "Fetching image for similarity map",
            extra={"image_url": request.image_url},
        )

        image_response = await asyncio.to_thread(
            requests.get, request.image_url, timeout=30
        )
        image_response.raise_for_status()
        image_bytes = image_response.content

        if not image_bytes:
            raise HTTPException(
                status_code=400,
                detail="Failed to fetch image: empty response",
            )

    except requests.exceptions.RequestException as e:
        logger.error(
            "Failed to fetch image",
            extra={"image_url": request.image_url, "error": str(e)},
        )
        raise HTTPException(
            status_code=400,
            detail=f"Failed to fetch image from URL: {str(e)}",
        )

    # Generate similarity maps
    try:
        logger.info(
            "Generating similarity maps",
            extra={
                "query": request.query[:50] + "..." if len(request.query) > 50 else request.query,
                "selected_tokens": request.selected_tokens,
                "alpha": request.alpha,
            },
        )

        result = await asyncio.to_thread(
            colpali_client.generate_similarity_maps,
            image_bytes=image_bytes,
            query=request.query,
            selected_tokens=request.selected_tokens,
            alpha=request.alpha,
        )

        # Parse response into typed models
        tokens = [
            TokenInfo(
                index=t["index"],
                token=t["token"],
                should_filter=t["should_filter"],
            )
            for t in result.get("tokens", [])
        ]

        similarity_maps = [
            SimilarityMapResult(
                token_index=m["token_index"],
                token=m["token"],
                similarity_map_base64=m["similarity_map_base64"],
            )
            for m in result.get("similarity_maps", [])
        ]

        logger.info(
            "Similarity maps generated successfully",
            extra={
                "query": request.query[:50] + "..." if len(request.query) > 50 else request.query,
                "token_count": len(tokens),
                "map_count": len(similarity_maps),
            },
        )

        return SimilarityMapResponse(
            query=result["query"],
            tokens=tokens,
            similarity_maps=similarity_maps,
        )

    except Exception as e:
        logger.exception("Failed to generate similarity maps")
        raise HTTPException(
            status_code=500,
            detail=f"Failed to generate similarity maps: {str(e)}",
        )


@router.get("/tokens")
async def get_query_tokens(query: str) -> List[TokenInfo]:
    """Get tokenized query with filter information.

    This is a lightweight endpoint to preview which tokens would be
    used for similarity map generation without actually generating maps.

    Args:
        query: The search query text

    Returns:
        List of TokenInfo with token strings and filter status
    """
    try:
        colpali_client = get_colpali_client()
    except Exception as e:
        logger.error("ColPali client not available", extra={"error": str(e)})
        raise HTTPException(
            status_code=503,
            detail="ColPali service unavailable",
        )

    # For token preview, we can use a dummy image or implement a separate endpoint
    # For now, return an error indicating this needs server-side implementation
    raise HTTPException(
        status_code=501,
        detail="Token preview endpoint not yet implemented. Use similarity-map endpoint directly.",
    )