# frontend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vision RAG API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/interpretability/similarity-map": {
      "post": {
        "tags": [
          "interpretability"
        ],
        "summary": "Generate Similarity Map",
        "description": "Generate interpretability similarity maps for query tokens on an image.\n\nThis endpoint creates visual overlays showing which parts of the image\nare most similar to each query token, enabling interpretability analysis.\n\nThe image is fetched from the provided URL (typically from MinIO storage),\nand similarity maps are generated by the ColPali service.\n\nArgs:\n    request: SimilarityMapRequest containing image_url, query, and optional parameters\n\nReturns:\n    SimilarityMapResponse with token info and base64-encoded similarity map images",
        "operationId": "generate_similarity_map_interpretability_similarity_map_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimilarityMapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarityMapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interpretability/tokens": {
      "get": {
        "tags": [
          "interpretability"
        ],
        "summary": "Get Query Tokens",
        "description": "Get tokenized query with filter information.\n\nThis is a lightweight endpoint to preview which tokens would be\nused for similarity map generation without actually generating maps.\n\nArgs:\n    query: The search query text\n\nReturns:\n    List of TokenInfo with token strings and filter status",
        "operationId": "get_query_tokens_interpretability_tokens_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TokenInfo"
                  },
                  "title": "Response Get Query Tokens Interpretability Tokens Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SimilarityMapRequest": {
        "properties": {
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "selected_tokens": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Tokens"
          },
          "alpha": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Alpha",
            "default": 0.5
          }
        },
        "type": "object",
        "required": [
          "image_url",
          "query"
        ],
        "title": "SimilarityMapRequest",
        "description": "Request for similarity map generation."
      },
      "SimilarityMapResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/TokenInfo"
            },
            "type": "array",
            "title": "Tokens"
          },
          "similarity_maps": {
            "items": {
              "$ref": "#/components/schemas/SimilarityMapResult"
            },
            "type": "array",
            "title": "Similarity Maps"
          }
        },
        "type": "object",
        "required": [
          "query",
          "tokens",
          "similarity_maps"
        ],
        "title": "SimilarityMapResponse",
        "description": "Response containing similarity maps."
      },
      "SimilarityMapResult": {
        "properties": {
          "token_index": {
            "type": "integer",
            "title": "Token Index"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "similarity_map_base64": {
            "type": "string",
            "title": "Similarity Map Base64"
          }
        },
        "type": "object",
        "required": [
          "token_index",
          "token",
          "similarity_map_base64"
        ],
        "title": "SimilarityMapResult",
        "description": "Similarity map for a single token."
      },
...
      "TokenInfo": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "should_filter": {
            "type": "boolean",
            "title": "Should Filter"
          }
        },
        "type": "object",
        "required": [
          "index",
          "token",
          "should_filter"
        ],
        "title": "TokenInfo",
        "description": "Information about a query token."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}